change all method and do exception

+++ src/main/java/step/managerPass/Service/ServiceEntityService.java
@@ -1,17 +1,22 @@
 package step.managerPass.Service;
 
+import com.sun.jdi.request.InvalidRequestStateException;
+import javassist.NotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.graalvm.compiler.core.common.type.ArithmeticOpTable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import step.managerPass.Entity.ServiceEntity;
+import step.managerPass.Entity.UserEntity;
 import step.managerPass.Repository.ServiceEntityRepository;
 
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,41 +24,45 @@ public class ServiceEntityService {
     private final ServiceEntityRepository serviceEntityRepository;
 
 
-    public ResponseEntity<List<ServiceEntity>> getAllServices() {
-        try {
+    public List<ServiceEntity> getAllServices() throws NotFoundException {
 
-            List<ServiceEntity> services = new ArrayList<>(serviceEntityRepository.findAll());
-            if (services.isEmpty()) {
-                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-            }
-            return new ResponseEntity<>(services, HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+        List<ServiceEntity> services = new ArrayList<>(serviceEntityRepository.findAll());
+
+        if (services.isEmpty()) {
+            throw new NotFoundException("Невозможно найти сервисы");
         }
+
+        return services;
     }
 
-    public ResponseEntity<ServiceEntity> addService(@RequestBody ServiceEntity newService) {
+    public ServiceEntity addService(ServiceEntity newService) {
+
         ServiceEntity service = serviceEntityRepository.save(newService);
-        return new ResponseEntity<>(service, HttpStatus.CREATED);
+        return service;
     }
 
-    public ResponseEntity<ServiceEntity> deleteService(@PathVariable Long id) {
+    public void deleteService(Long id) throws NotFoundException {
+        Optional<ServiceEntity> optionalServiceEntity = serviceEntityRepository.findById(id);
+        if (optionalServiceEntity.isPresent()) {
+            throw new NotFoundException("Невозможно найти сервис" + id + "с таким идентификатором");
+        }
         serviceEntityRepository.deleteById(id);
-        return new ResponseEntity<>(null, HttpStatus.OK);
     }
 
-    public ResponseEntity<ServiceEntity> updateService(@RequestBody ServiceEntity newService, @PathVariable Long id) {
-        serviceEntityRepository.findById(id).map(service -> {
-            service.setName(newService.getName());
-            service.setId_service(newService.getId_service());
-            serviceEntityRepository.save(service);
-            return new ResponseEntity<>(null, HttpStatus.OK);
-        })
-                .orElseGet(() -> {
-                    newService.setId_service(id);
-                    serviceEntityRepository.save(newService);
-                    return new ResponseEntity(null, HttpStatus.CREATED);
-                });
-        return new ResponseEntity<>(null, HttpStatus.OK);
+    public ServiceEntity updateService(ServiceEntity serviceEntity) throws NotFoundException {
+
+        if (serviceEntity == null || serviceEntity.getId_service() == null) {
+            throw new InvalidRequestStateException("Объект или идентификатор не может быть пустым");
+        }
+
+        Optional<ServiceEntity> optionalServiceEntity = serviceEntityRepository.findById(serviceEntity.getId_service());
+        if (optionalServiceEntity.isPresent()) {
+            throw new NotFoundException("Невозможно найти сервис" + serviceEntity.getId_service() + "с таким идентификатором");
+        }
+        ServiceEntity existingServiceEntity = optionalServiceEntity.get();
+        existingServiceEntity.setName(serviceEntity.getName());
+        existingServiceEntity.setComment(serviceEntity.getComment());
+
+        return serviceEntityRepository.save(existingServiceEntity);
     }
 }
